Update 500.es.adoc
Corrected references for "WORMS" to "WoRMS", from "Github" to "GitHub", and from "Species matching" to "Species Matching". Added a missing space before a parenthesis. Corrected misspelled "construido". Replaced a . with a , in the middle of a sentence. Replaced "http://api.geonames.org/srtm1'" with "grel:\"http://api.geonames.org/srtm1" to avoid a non-existent URL link and to remove the spurious single quote at the end. Corrected "msnm" to "m" for differences in elevation. An elevation may be msnm, but a difference in elevation is in meters. Corrected typo "el elementos" to "el elemento".